refactor: addresses linting issues
refactor: move multiuse functions to dedicated_utils dir

refactor: addition of the clusterId to the cluster name during cluster selection prompt

refactor: addition of cluster id to cluster selection prompt for deregister cluster

refactor: addition of successful registration prompt
