+++ pkg/cmd/kafka/create/create.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	v1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/dedicated/dedicatedcmdutil"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/connection/api/clustermgmt"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/kafkautil"
 	"os"
@@ -465,7 +466,7 @@ func setEnterpriseClusterList(opts *options) (*kafkamgmtclient.EnterpriseCluster
 			return &kafkamgmtclient.EnterpriseClusterList{}, &emptyClusterMap, nil
 		}
 
-		return nil, nil, fmt.Errorf("%v, %v", response.Status, err)
+		return nil, nil, fmt.Errorf("%v, %w", response.Status, err)
 	}
 
 	clusterMap, err := getClusterNameMap(opts.f, kfmClusterList)
@@ -489,7 +490,10 @@ func selectEnterpriseOrRHInfraPrompt(opts *options) error {
 	}
 
 	var idx int
-	survey.AskOne(promptForCluster, &idx)
+	err := survey.AskOne(promptForCluster, &idx)
+	if err != nil {
+		return err
+	}
 	if idx == 0 {
 		opts.useEnterpriseCluster = false
 	} else {
@@ -578,7 +582,7 @@ func promptKafkaPayload(opts *options, constants *remote.DynamicServiceConstants
 
 		enterpriseQuota = orgQuota.EnterpriseQuotas[0]
 
-		// there is no quota left to use enteprise
+		// there is no quota left to use enterprise
 		if enterpriseQuota.Quota == 0 {
 			return nil, opts.f.Localizer.MustLocalizeError("kafka.create.error.noQuotaLeft")
 		}
@@ -608,7 +612,7 @@ func promptKafkaPayload(opts *options, constants *remote.DynamicServiceConstants
 
 	// if billing model is marketplace, prompt for marketplace provider and account id
 	if answers.BillingModel == accountmgmtutil.QuotaMarketplaceType {
-		answers, err = marketplaceQuotaPrompt(orgQuota, answers, f, err)
+		answers, err = marketplaceQuotaPrompt(orgQuota, answers, f)
 		if err != nil {
 			return nil, err
 		}
@@ -751,7 +755,7 @@ func promptKafkaPayload(opts *options, constants *remote.DynamicServiceConstants
 	return payload, nil
 }
 
-func marketplaceQuotaPrompt(orgQuota *accountmgmtutil.OrgQuotas, answers *promptAnswers, f *factory.Factory, err error) (*promptAnswers, error) {
+func marketplaceQuotaPrompt(orgQuota *accountmgmtutil.OrgQuotas, answers *promptAnswers, f *factory.Factory) (*promptAnswers, error) {
 	validMarketPlaces := FetchValidMarketplaces(orgQuota.MarketplaceQuotas, answers.CloudProvider)
 	if len(validMarketPlaces) == 1 {
 		answers.Marketplace = validMarketPlaces[0]
@@ -760,7 +764,7 @@ func marketplaceQuotaPrompt(orgQuota *accountmgmtutil.OrgQuotas, answers *prompt
 			Message: f.Localizer.MustLocalize("kafka.create.input.marketplace.message"),
 			Options: validMarketPlaces,
 		}
-		err = survey.AskOne(marketplacePrompt, &answers.Marketplace)
+		err := survey.AskOne(marketplacePrompt, &answers.Marketplace)
 		if err != nil {
 			return nil, err
 		}
@@ -777,7 +781,7 @@ func marketplaceQuotaPrompt(orgQuota *accountmgmtutil.OrgQuotas, answers *prompt
 				Message: f.Localizer.MustLocalize("kafka.create.input.marketplaceAccountID.message"),
 				Options: validMarketplaceAcctIDs,
 			}
-			err = survey.AskOne(marketplaceAccountPrompt, &answers.MarketplaceAcctID)
+			err := survey.AskOne(marketplaceAccountPrompt, &answers.MarketplaceAcctID)
 			if err != nil {
 				return nil, err
 			}
@@ -825,32 +829,8 @@ func getClusterNameMap(f *factory.Factory, clusterList *kafkamgmtclient.Enterpri
 
 }
 
-func validateClusters(clusterList *kafkamgmtclient.EnterpriseClusterList) *kafkamgmtclient.EnterpriseClusterList {
-	// if cluster is in a ready state add it to the list of clusters
-	items := make([]kafkamgmtclient.EnterpriseClusterListItem, 0, len(clusterList.Items))
-	for _, cluster := range clusterList.Items {
-		if *cluster.Status == "ready" {
-			items = append(items, cluster)
-		}
-	}
-
-	newClusterList := kafkamgmtclient.NewEnterpriseClusterList(clusterList.Kind, clusterList.Page, int32(len(items)), clusterList.Total, items)
-	return newClusterList
-}
-
-func createPromptOptionsFromClusters(clusterList *kafkamgmtclient.EnterpriseClusterList, clusterMap *map[string]v1.Cluster) *[]string {
-	promptOptions := []string{}
-	validatedClusters := validateClusters(clusterList)
-	for _, cluster := range validatedClusters.Items {
-		ocmCluster := (*clusterMap)[cluster.GetId()]
-		display := ocmCluster.Name() + " (" + cluster.GetId() + ")"
-		promptOptions = append(promptOptions, display)
-	}
-	return &promptOptions
-}
-
 func selectClusterPrompt(opts *options) (int, error) {
-	promptOptions := createPromptOptionsFromClusters(opts.kfmClusterList, opts.clusterMap)
+	promptOptions := dedicatedcmdutil.CreatePromptOptionsFromClusters(opts.kfmClusterList, opts.clusterMap)
 	promptForCluster := &survey.Select{
 		Message: opts.f.Localizer.MustLocalize("kafka.create.input.cluster.selectClusterMessage"),
 		Options: *promptOptions,
@@ -912,6 +892,7 @@ func getClusterDetails(f *factory.Factory, clusterId string) (*kafkamgmtclient.E
 
 	return &cluster, nil
 }
+
 func CreateNullableString(str *string) kafkamgmtclient.NullableString {
 	return *kafkamgmtclient.NewNullableString(str)
 }
